Former-commit-id: 4d0f24dd053f08551c0b18629c75cfef4d1e33ef

+++ arl_dev/src/CCA/Components/MD/Nonbonded/TwoBodyDeterministic.h
@@ -97,7 +97,7 @@ namespace Uintah {
     private:
       static const std::string nonbondedType;
       const double d_nonbondedRadius;
-      const double d_nonbondedGhostCells;
+      const int d_nonbondedGhostCells;
 
   };
 }
